Rename
- Rename files to be Sweet instead of Sweetness
- Move UITableView and UICollectionView extensions to their own files
